+++ Felix/view_state_fake.h
@@ -17,7 +17,7 @@ public:
 		SENSE("show_content") ;
 
 	}
-	void retrieve_edit_record(int mem_id, mem_engine::record_pointer new_rec, bool) 
+	void retrieve_edit_record(size_t mem_id, mem_engine::record_pointer new_rec, bool) 
 	{
 		new_rec ;
 		SENSE("retrieve_edit_record") ;
