+++ Felix/view_state_initial.cpp
@@ -12,7 +12,7 @@ void ViewStateInitial::handle_toggle_edit_mode()
 	m_view->put_edit_mode( ! m_view->is_edit_mode() ) ;
 }
 
-void ViewStateInitial::retrieve_edit_record(int mem_id, mem_engine::record_pointer new_rec, bool is_add)
+void ViewStateInitial::retrieve_edit_record(size_t mem_id, mem_engine::record_pointer new_rec, bool is_add)
 {
 	mem_engine::memory_pointer mem = m_model->get_memory_by_id(mem_id) ;
 	if (is_add)
