+++ Felix/view_state_review.h
@@ -21,7 +21,7 @@ public:
 	void set_current(size_t num);
 	virtual size_t get_current();
 	void handle_toggle_edit_mode();
-	void retrieve_edit_record(int mem_id, mem_engine::record_pointer new_rec, bool is_add);
+	void retrieve_edit_record(size_t mem_id, mem_engine::record_pointer new_rec, bool is_add);
 	void show_content();
 	mem_engine::search_match_ptr get_current_match();
 	void activate();
